refactor: replace SimpleStudentTreeProvider with proper StudentCourseContentTreeProvider
